Changed Int64 to Int

+++ src/sensitivities/array.jl
@@ -5,7 +5,7 @@ Base.size(x::Node, dims...) = size(x.val, dims...)
 Base.length(x::Node) = length(x.val)
 
 # Sensitivity for the first argument of `reshape`.
-@explicit_intercepts reshape Tuple{∇RealArray, Vararg{Int64}} [true, false]
-@explicit_intercepts reshape Tuple{∇RealArray, Tuple{Vararg{Int64}}} [true, false]
+@explicit_intercepts reshape Tuple{∇RealArray, Vararg{Int}} [true, false]
+@explicit_intercepts reshape Tuple{∇RealArray, Tuple{Vararg{Int}}} [true, false]
 ∇(::typeof(reshape), ::Type{Arg{1}}, _, y, ȳ, A::∇RealArray, args...) =
     reshape(ȳ, size(A)...)
